src/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-annoying==0.10.6
django-filter==2.4.0
django-jsonstore==0.5.0
django-material==1.7.4
django-viewflow==1.8.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
